+++ src/WriterModule/f142_test/f142_TestWriter.cpp
@@ -14,14 +14,10 @@
 namespace WriterModule {
 namespace f142 {
 
-/// Parse the configuration for this stream.
-void f142_TestWriter::parse_config(std::string const &) {}
-
 /// \brief Implement the writer module interface, forward to the CREATE case
 /// of
 /// `init_hdf`.
-InitResult f142_TestWriter::init_hdf(hdf5::node::Group &HDFGroup,
-                                     std::string const &) {
+InitResult f142_TestWriter::init_hdf(hdf5::node::Group &HDFGroup) {
   auto Create = NeXusDataset::Mode::Create;
   try {
     NeXusDataset::CueIndex(HDFGroup, Create,