+++ src/WriterModule/f142_test/f142_TestWriter.h
@@ -20,17 +20,14 @@ using FlatbufferMessage = FileWriter::FlatbufferMessage;
 class f142_TestWriter : public WriterModule::Base {
 public:
   /// Implements writer module interface.
-  InitResult init_hdf(hdf5::node::Group &HDFGroup,
-                      std::string const &HDFAttributes) override;
-  /// Implements writer module interface.
-  void parse_config(std::string const &ConfigurationStream) override;
+  InitResult init_hdf(hdf5::node::Group &HDFGroup) override;
   /// Implements writer module interface.
   WriterModule::InitResult reopen(hdf5::node::Group &HDFGroup) override;
 
   /// Write an incoming message which should contain a flatbuffer.
   void write(FlatbufferMessage const &Message) override;
 
-  f142_TestWriter() : WriterModule::Base(false) {}
+  f142_TestWriter() : WriterModule::Base(false, "NXlog") {}
   ~f142_TestWriter() override = default;
 
 protected: