+++ src/WriterModule/ns10/ns10_Writer.h
@@ -17,33 +17,30 @@
 #include "FlatbufferMessage.h"
 #include "NeXusDataset/NeXusDataset.h"
 #include "WriterModuleBase.h"
+#include "WriterModuleConfig/Field.h"
 
 namespace WriterModule {
 namespace ns10 {
 
 class ns10_Writer : public WriterModule::Base {
 public:
-  ns10_Writer() : WriterModule::Base(false) {}
+  ns10_Writer() : WriterModule::Base(false, "NXlog") {}
   ~ns10_Writer() override = default;
 
-  void parse_config(std::string const &ConfigurationStream) override;
-
-  InitResult init_hdf(hdf5::node::Group &HDFGroup,
-                      std::string const &HDFAttributes) override;
+  InitResult init_hdf(hdf5::node::Group &HDFGroup) override;
 
   InitResult reopen(hdf5::node::Group &HDFGroup) override;
 
   void write(FileWriter::FlatbufferMessage const &Message) override;
 
 protected:
-  std::string Sourcename;
   NeXusDataset::DoubleValue Values;
-  hdf5::Dimensions ChunkSize{1024};
   NeXusDataset::Time Timestamp;
-  int CueInterval{1000};
   int CueCounter{0};
   NeXusDataset::CueIndex CueTimestampIndex;
   NeXusDataset::CueTimestampZero CueTimestamp;
+  WriterModuleConfig::Field<int> CueInterval{this, "cue_interval", 1000};
+  WriterModuleConfig::Field<size_t> ChunkSize{this, "chunk_size", 1024};
 
 private:
   SharedLogger Logger = spdlog::get("filewriterlogger");