+++ src/WriterModule/senv/senv_Writer.h
@@ -19,6 +19,7 @@
 #include "Msg.h"
 #include "NeXusDataset/NeXusDataset.h"
 #include "WriterModuleBase.h"
+#include "WriterModuleConfig/Field.h"
 
 namespace WriterModule {
 namespace senv {
@@ -32,13 +33,10 @@ std::vector<std::uint64_t> GenerateTimeStamps(std::uint64_t OriginTimeStamp,
 /// See parent class for documentation.
 class senv_Writer : public FileWriterBase {
 public:
-  senv_Writer() : FileWriterBase(false) {}
+  senv_Writer() : FileWriterBase(false, "NXlog") {}
   ~senv_Writer() override = default;
 
-  void parse_config(std::string const &) override;
-
-  InitResult init_hdf(hdf5::node::Group &HDFGroup,
-                      std::string const &HDFAttributes) override;
+  InitResult init_hdf(hdf5::node::Group &HDFGroup) override;
 
   InitResult reopen(hdf5::node::Group &HDFGroup) override;
 
@@ -50,6 +48,7 @@ protected:
   NeXusDataset::CueIndex CueTimestampIndex;
   NeXusDataset::CueTimestampZero CueTimestamp;
   SharedLogger Logger = spdlog::get("filewriterlogger");
+  WriterModuleConfig::Field<size_t> ChunkSize{this, "chunk_size", 4096};
 };
 } // namespace senv
 } // namespace WriterModule