+++ src/tests/Stream/MessageWriterTests.cpp
@@ -17,11 +17,9 @@
 
 class WriterModuleStandIn : public WriterModule::Base {
 public:
-  WriterModuleStandIn() : WriterModule::Base(true) {}
-  MAKE_MOCK1(parse_config, void(std::string const &), override);
-  MAKE_MOCK2(init_hdf,
-             WriterModule::InitResult(hdf5::node::Group &, std::string const &),
-             override);
+  MAKE_MOCK0(config_post_processing, void(), override);
+  WriterModuleStandIn() : WriterModule::Base(true, "test") {}
+  MAKE_MOCK1(init_hdf, WriterModule::InitResult(hdf5::node::Group &), override);
   MAKE_MOCK1(reopen, WriterModule::InitResult(hdf5::node::Group &), override);
   MAKE_MOCK1(write, void(FileWriter::FlatbufferMessage const &), override);
 };