+++ src/tests/helpers/StubWriterModule.h
@@ -15,10 +15,8 @@ using WriterModule::InitResult;
 
 class StubWriterModule : public WriterModule::Base {
 public:
-  StubWriterModule() : WriterModule::Base(false) {}
-  void parse_config(std::string const & /*ConfigurationStream*/) override {}
-  InitResult init_hdf(hdf5::node::Group & /*HDFGroup*/,
-                      std::string const & /*HDFAttributes*/) override {
+  StubWriterModule() : WriterModule::Base(false, "NXlog") {}
+  InitResult init_hdf(hdf5::node::Group & /*HDFGroup*/) override {
     return InitResult::OK;
   }
   InitResult reopen(hdf5::node::Group & /*HDFGroup*/) override {